+++ src/main/kotlin/com/coder/toolbox/CoderToolboxContext.kt
@@ -5,6 +5,7 @@ import com.coder.toolbox.store.CoderSecretsStore
 import com.coder.toolbox.store.CoderSettingsStore
 import com.coder.toolbox.util.toURL
 import com.jetbrains.toolbox.api.core.diagnostics.Logger
+import com.jetbrains.toolbox.api.core.os.LocalDesktopManager
 import com.jetbrains.toolbox.api.localization.LocalizableStringFactory
 import com.jetbrains.toolbox.api.remoteDev.connection.ClientHelper
 import com.jetbrains.toolbox.api.remoteDev.connection.ToolboxProxySettings
@@ -18,6 +19,7 @@ data class CoderToolboxContext(
     val envPageManager: EnvironmentUiPageManager,
     val envStateColorPalette: EnvironmentStateColorPalette,
     val ideOrchestrator: ClientHelper,
+    val desktop: LocalDesktopManager,
     val cs: CoroutineScope,
     val logger: Logger,
     val i18n: LocalizableStringFactory,
@@ -62,5 +64,4 @@ data class CoderToolboxContext(
             } else null
         }
     }
-
 }