+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - render network status in the Settings tab, under `Additional environment information` section.
+- quick action for creating new workspaces from the web dashboard.
 
 ## 0.2.1 - 2025-05-05